requirements.txt
Flask
gunicorn
sqlalchemy
psycopg2
python-dotenv
mailjet_rest
werkzeug